fix: run playground on Windows
Yalc uses the `LOCALAPPDATA` environment variable to determine where to put local packages.
Without this change, turbo does not propagate `LOCALAPPDATA` to the `build` commands in packages. The packages get published to ``%HOME%\.yalc` instead of `%LOCALAPPDATA%\Yalc`. This means that `playground:publish` will not find them in `%LOCALAPPDATA%\Yalc`.